requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements.in'
#
allure-pytest==2.15.0
    # via -r requirements.in
allure-python-commons==2.15.0
    # via allure-pytest
appium-python-client==5.2.4
    # via -r requirements.in
attrs==25.4.0
    # via
    #   allure-python-commons
    #   outcome
    #   trio
beautifulsoup4==4.14.2
    # via -r requirements.in
certifi==2025.11.12
    # via
    #   requests
    #   selenium
cffi==2.0.0
    # via trio
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via pytest
comtypes==1.4.13
    # via pywinauto
execnet==2.1.2
    # via pytest-xdist
h11==0.16.0
    # via wsproto
idna==3.11
    # via
    #   requests
    #   trio
iniconfig==2.3.0
    # via pytest
mouseinfo==0.1.3
    # via pyautogui
numpy==2.3.5
    # via pandas
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via pytest
pandas==2.3.3
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
psutil==7.1.3
    # via -r requirements.in
pyautogui==0.9.54
    # via -r requirements.in
pycparser==2.23
    # via cffi
pygetwindow==0.0.9
    # via pyautogui
pygments==2.19.2
    # via pytest
pymsgbox==2.0.1
    # via pyautogui
pyperclip==1.11.0
    # via mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
pysocks==1.7.1
    # via urllib3
pytest==9.0.1
    # via
    #   -r requirements.in
    #   allure-pytest
    #   pytest-xdist
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytweening==1.2.0
    # via pyautogui
pytz==2025.2
    # via pandas
pywin32==311
    # via pywinauto
pywinauto==0.6.9
    # via -r requirements.in
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
selenium==4.38.0
    # via
    #   -r requirements.in
    #   appium-python-client
six==1.17.0
    # via
    #   python-dateutil
    #   pywinauto
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.8
    # via beautifulsoup4
trio==0.32.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.15.0
    # via
    #   appium-python-client
    #   beautifulsoup4
    #   selenium
tzdata==2025.2
    # via pandas
urllib3[socks]==2.5.0
    # via
    #   requests
    #   selenium
websocket-client==1.9.0
    # via selenium
wsproto==1.3.1
    # via trio-websocket
